requirements.txt
torch==2.3.1
torchvision==0.18.1
matplotlib==3.9.0
kaggle==1.6.14
Pillow==10.3.0
tqdm==4.66.4
Flask==3.0.3